remove wrong placed tram shelter
